% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{All_cnts}
\alias{All_cnts}
\title{Visual counts of Kokanee spawners in British COlumbia and Pink Salmon spawners in Alaska.}
\format{
A data frame with 145 rows and 9 variables:
\describe{
  \item{SPECIES}{Character string describing common name of species counted}
  \item{LAKE}{Character string describing the rearing water body where spawners originate}
  \item{STREAM}{Character string describing stream name}
  \item{REACH}{Character string describing stream reaches counted}
  \item{DATE}{Numerical date}
  \item{YEAR}{Year of survey}
  \item{MONTH}{Month of survey}
  \item{DAY}{Day within month of survey}
  \item{GROUND_LIVE}{Number of live spawners counted in stream from an observer on the ground}
  \item{GROUND_DEAD}{Number of dead spawners counted in stream from an observer on the ground}
  \item{GROUND_NEW_DEAD}{Number of newly dead spawners counted in stream from an observer on the ground. KNown to be new if the observer cuts tail off of carcasses after observing.}
  \item{TEMP_C}{Stream temperature measured at time of survey or daily average by a logger}
  \item{COMMENTS}{COmments related to data}
  \item{FENCE_PASS}{The number of spawners moving past a counting fence}
  \item{AIR_LIVE}{Number of live spawners counted in stream from an observer in the air - helicopter}
  \item{GROUND_INTERP}{0 or 1 indicating if the GROUND_LIVE data was interpolated for that day}
  \item{FENCE_INTERP}{0 or 1 indicating if the FENCE_PASS live data was interpolated for that day}
  \item{ADDED_0}{0 or 1 indicating if a 0 observation was added (assumed) on that day to define the spawning period for AUC}
  \item{DOY}{Day within year (from Jan 1) of survey}
  \item{STREAM_YR}{Grouping variable for stream and year}
  \item{KO_sim}{0 or 1 to Subsample of counts to simulate ground counts following Kokanee methods}

}
}
\source{
Data supplied from R. Millar for Pink salmon and BC government staff in Penticton for Kokanee. PAskey personal copies with .csv files stored in data-raw. data_prep() function in data-raw combines SPAWNER_COUNTS.csv, Added_)s.cvs, and filters and formats to generate .rda file.
}
\usage{
All_cnts
}
\description{
Visual counts of Kokanee spawners in British COlumbia and Pink Salmon spawners in Alaska.
}
\keyword{datasets}
