% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAUC.R
\name{GAUC}
\alias{GAUC}
\title{GAUC}
\usage{
GAUC(t, obs, day1 = NULL, lday = NULL, out = "AUC")
}
\arguments{
\item{t}{an ordered time series of count days}

\item{obs}{a series of observed counts of live fish corresponding to each t}

\item{day1}{first day to sum in fish days, an optional parameter if the user wishes to restrict or extend the predictions or summed fish days to a specific period}

\item{lday}{last day to sum in fish days, an optional parameter if the user wishes to restrict or extend the predictions or summed fish days to a specific period}

\item{out}{preferred type of output, default "AUC" = total fish day, "preds" returns the time specific predictions (e.g. for plotting)}
}
\description{
A function to estimate the GAUC from a time series of spawner counts
}
\details{
This function is used to calculate the number of total fish days based on a series of spawner counts. Assumes the relative number of fish over time follows a Gaussian curve.
}
\examples{
 #Kokanee data from Coldstream Creek in 2021
t = c(270,273,276,282,287,290,295)
obs = c(603,1010,1380,4183,4158,3721,4423)
 #Add 0 observations at start and end, so that same information as used for TAUC example
 #Although 0 observations are not necessary for GAUC
t0 = c(263,303)
o0 = c(0,0)
t = c(t0,t)
obs = c(o0,obs)
GAUC(t,obs)
}
\keyword{AUC}
\keyword{GAUC}
