% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{KO_cnts}
\alias{KO_cnts}
\title{Visual counts of Kokanee spawners in Okanagan, BC streams from ground based surveys}
\format{
A data frame with 145 rows and 9 variables:
\describe{
  \item{STREAM}{Character string describing stream name}
  \item{YEAR}{Year of survey}
  \item{MONTH}{Month of survey}
  \item{DAY}{Day within month of survey}
  \item{DOY}{Day within year (from Jan 1) of survey}
  \item{Date}{Character string of date}
  \item{NUM_DATE}{Numeric of date}
  \item{LIVE_COUNT}{Number of live Kokanee counted in stream}
  \item{SOURCE}{Character string describing if 'OBSERVED' data or an "ADDED_0" to define endpoints of run for AUC calculations}
}
}
\source{
PAskey personal copy with .csv stored in data-raw. Obtained from BC provincial government fisheries staff in Penticton.
}
\usage{
KO_cnts
}
\description{
Visual counts of Kokanee spawners in Okanagan, BC streams from ground based surveys
}
\keyword{datasets}
