% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{KO_fence}
\alias{KO_fence}
\title{Fence counts of all Kokanee entering stream over entire run for several BC interior streams}
\format{
A data frame with 14 rows and 3 variables:
\describe{
  \item{STREAM}{Character string describing stream name}
  \item{YEAR}{Year of survey}
  \item{FENCE_COUNT}{Total number of fish passing fence in the survey year}
}
}
\source{
PAskey personal copy with .csv stored in data-raw
}
\usage{
KO_fence
}
\description{
Fence counts of all Kokanee entering stream over entire run for several BC interior streams
}
\keyword{datasets}
