% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{Mission_ch}
\alias{Mission_ch}
\title{Daily visual counts of Kokanee spawners in Mission Creek spawning channel, BC from ground based surveys}
\format{
A data frame with 616 observations and 10 variables
\describe{
  \item{STREAM}{Character string describing stream name, only MISSION}
  \item{REACH}{Character string describing stream reach. Reach E selected as this is the spawning channel}
  \item{YEAR}{Year of survey}
  \item{MONTH}{Month of survey}
  \item{DAY}{Day within month of survey}
  \item{Date}{Character string of date}
  \item{NO_LIVE}{Number of live Kokanee counted in stream}
   \item{NO_DEAD}{Number of dead Kokanee counted in stream}
  \item{TEMP_C}{Numeric of stream temperature in celsius}
  \item{COMMENTS}{Character string of additional information collected by counting contractor}
}
}
\source{
PAskey personal copy. Raw csv file Mission_channel in data-raw folder. Obtained from BC fisheries staff in Penticton.
}
\usage{
Mission_ch
}
\description{
Daily visual counts of Kokanee spawners in Mission Creek spawning channel, BC from ground based surveys
}
\keyword{datasets}
