% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Repro_tbls.R
\name{Repro_tbls}
\alias{Repro_tbls}
\title{Repro_tbls}
\usage{
Repro_tbls()
}
\arguments{
\item{Output}{a character vector to describe output?}
}
\description{
A function to reproduce all results in tables in Askey et al. paper
}
\keyword{Askey}
\keyword{Kokanee,}
\keyword{Pink,}
\keyword{spwnr,}
