% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{spwnr_ests}
\alias{spwnr_ests}
\title{A summary of the total number of spawners passing through a counting fence as compared to visual indices of fish days (GAUC and TAUC) and Peak Counts.}
\format{
A data frame with 25 rows and 8 variables:
\describe{
  \item{STREAM_YR}{Grouping variable for stream and year}
  \item{SPECIES}{Character string describing species of fish being counted, Kokanee or Pink Salmon}
  \item{STREAM}{Character string describing stream name}
  \item{YEAR}{Year of survey}
  \item{Method}{The viewing method GROUND or AIR used to count fish}
  \item{FENCE}{Total number of fish passing fence in the survey year}
  \item{PEAK_COUNT}{Maximum number of fish observed during a single survey for the corresponding species, year and method}
  \item{TAUC}{Total fish days from Trapezoidal area under the curve}
  \item{GAUC}{Total fish days from Gaussian area under the curve}
  \item{fit_group}{Groups of streams for cross-validation analyses}

}
}
\source{
Data supplied from R. Millar for Pink salmon and BC government staff in Penticton for Kokanee. PAskey personal copy compiled from .csv files in data-raw. See data_prep.R in data-raw.
}
\usage{
spwnr_ests
}
\description{
A summary of the total number of spawners passing through a counting fence as compared to visual indices of fish days (GAUC and TAUC) and Peak Counts.
}
\keyword{datasets}
